requirements.txt
asgiref==3.7.2
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.5
cssselect2==0.7.0
defusedxml==0.8.0rc2
asgiref==3.7.2
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.5
cssselect2==0.7.0
defusedxml==0.8.0rc2
Django==3.2.3
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
fonttools==4.50.0
gunicorn==20.1.0
html5lib==1.1
idna==3.6
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydyf==0.9.0
PyJWT==2.8.0
pyphen==0.14.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tinycss2==1.2.1
toml==0.10.2
typing_extensions==4.10.0
urllib3==2.2.1
webencodings==0.5.1
zopfli==0.2.3
